requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
clearcache-python==0.0.8
Django==3.1.1
django-appconf==1.0.4
django-bleach==0.6.1
django-compressor==2.4
django-filter==2.3.0
django-generic-json-views==0.8
django-taggit==1.3.0
django-widget-tweaks==1.4.8
gunicorn==20.0.4
html2text==2020.1.16
html5lib==1.1
idna==2.10
lorem==0.1.1
mysqlclient==2.0.1
packaging==20.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-ldap==3.3.1
pytz==2020.1
rcssmin==1.0.6
requests==2.24.0
rjsmin==1.1.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.10
webencodings==0.5.1
